refactor: extract logic for generating es query from needle filter query
